Improve log output
When the `deploy key` method is used, instead of the `ssh`
warning message, an empty line is added before the output of the
`commit and push` step.

  * https://cloud.githubusercontent.com/assets/1223565/13126766/80981da6-d5d4-11e5-909d-8c22d2a01da6.png
  * https://cloud.githubusercontent.com/assets/1223565/13126767/809afcec-d5d4-11e5-8343-908870a34f0d.png

That happens because the warning message generated by `ssh` on
`git push` ends with also a carriage return character (`\r`).

  " Warning: Permanently added the RSA host key for IP address
    '192.30.252.128' to the list of known hosts. "

The carriage return character (`\r`) is interpreted¹ by `printf`,
moves the active position to the beginning of the line, and then,
the newline character (`\n`) basically overwrites the entire line,
thus, outputting an empty line instead of the warning message.

- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -

This commit improves the log output by:

  * fixing issues such as the one described above
  * redirecting the output of all descriptors² through
    the `remove_sensitive_information` function

- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -

¹ http://ss64.com/bash/printf.html
² http://www.tldp.org/LDP/abs/html/io-redirection.html