Use a single-char for short command-line options
The general convention¹ is for single-character (short) options to
use `-` and multi-character (long) options to use `--`.

In order to match the command line options of certain APIs such as
OpenSSL's², some of the travis-scripts used multi-character options
with `-`.

This commit changes those scripts so that they conform with / follow
the general convention.

- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -

¹ http://www.faqs.org/docs/artu/ch10s05.html

² From https://www.openssl.org/docs/manmaster/apps/enc.html:

  " -iv IV

        the actual IV to use: this must be represented as a string
        comprised only of hex digits. When only the key is specified
        using the -K option, the IV must explicitly be defined.
        When a password is being specified using one of the other
        options, the IV is generated from this password. "

- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -

Fix alrra/travis-scripts#23