Remove support for GitHub access tokens
GitHub access tokens are like passwords. They basically grant access
not only to the repository in which they are used, but to everything
the user that created them has access to.

This makes them less secure then using SSH keys, which, as opposed
to access tokens, only grant access to the repository in which they
are used.

- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -

See also: https://developer.github.com/guides/managing-deploy-keys/

Close alrra/travis-scripts#25