Allow `master` as the distribution branch
The `update-branch.sh` script fails when `master` is specified as
the distribution branch. That happens because Git will automatically
create the (default) `master` branch when the commit is made, thus,
preventing the script from creating that branch:

    " fatal: A branch named 'master' already exists. "

This commit fixes the issue described above by introducing an extra
check to ensure that the `update-branch.sh` script doesn't try to
create the `master` branch again.

- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -

Fix alrra/travis-scripts#22